65.1.A.5.65: Open source code for Xperia 10 IV
Open source code for Xperia 10 IV (XQ-CC54, XQ-CC72); software version 65.1.A.5.65

+++ drivers/gpu/drm/i915/gem/i915_gem_execbuffer.c
@@ -3,6 +3,11 @@
  *
  * Copyright © 2008,2010 Intel Corporation
  */
+/*
+ * NOTE: This file has been modified by Sony Corporation.
+ * Modifications are Copyright 2021 Sony Corporation,
+ * and licensed under the license of the file.
+ */
 
 #include <linux/intel-iommu.h>
 #include <linux/dma-resv.h>
@@ -1452,7 +1457,7 @@ static int eb_relocate_vma(struct i915_execbuffer *eb, struct i915_vma *vma)
 
 	urelocs = u64_to_user_ptr(entry->relocs_ptr);
 	remain = entry->relocation_count;
-	if (unlikely((unsigned long)remain > N_RELOC(ULONG_MAX)))
+	if (unlikely(remain > N_RELOC(ULONG_MAX)))
 		return -EINVAL;
 
 	/*
